Crate signal_neon_futures

Allows async blocks to be used to wait on JavaScript futures using Neon.

Neon provides a way to expose synchronous JavaScript functions from Rust. This means that if Rust wants to wait for the result of a JavaScript promise, it can at best return a callback to continue its work when the promise settles. This does not naturally compose with Rust's async, which works in terms of Futures.

This crate provides functionality for (1) wrapping JavaScript futures so they can be awaited on in Rust, and (2) producing a JavaScript promise that wraps a Rust future. It does so by resuming execution of the Rust future on the JavaScript microtask queue whenever an awaited JavaScript promise is settled.

To get started, look at the promise() function and the JsFuture::from_promise method.

Structs

JsFuture
A future representing the result of a JavaScript promise.

JsFutureBuilder
Sets up a JsFuture using a builder pattern. See JsFuture::get_promise.

PersistentException
Persists a JavaScript exception across call contexts for later re-throwing (or promise rejection).

Traits

ContextEx
Adds support for executing closures and futures on the JavaScript main thread's microtask queue.

EventQueueEx
Adds support for executing futures on a Neon EventQueue.

Functions

call_method
A convenience for calling a method on an object.

promise
Produces a JavaScript Promise that represents the result of a Rust computation.

settle_promise
Use this to return your "settle" function when using promise().

Type Definitions

JsPromiseResult
The result of a JavaScript promise: a success value or a failure value.