Function signal_neon_futures::promise

pub fn promise<'a, V, F, R>(
    cx: &mut FunctionContext<'a>, 
    callback: impl FnOnce(&mut FunctionContext<'a>, JsAsyncContext) -> R
) -> JsResult<'a, JsObject> where
    V: Value,
    F: for<'b> FnOnce(&mut FunctionContext<'b>) -> JsResult<'b, V>,
    R: Future<Output = Result<F, JsAsyncContextKey<JsValue>>> + 'static, 

Produces a JavaScript Promise that represents the result of a Rust computation.

There's a lot going on here, so let's break it down:

callback is a function that is called synchronously to produce a future (async block).

That resulting future (R) must itself return a "fulfill" function, or fail with a value stored in the JsAsyncContext.

The "fulfill" function (F) must produce a result in a synchronous JavaScript context, or fail with a JavaScript exception.

That value (V) will be the final result of the JavaScript Promise.

If there are any failures during the evaluation of the future or "fulfill" function, they will result in the rejection of the Promise.

If there are any panics during the evaluation of the future or "fulfill" function, they will be translated to JavaScript Errors (per Neon conventions).

In practice, this is going to be the easiest way to produce a JavaScript Promise, especially for safely handling errors.

fn js_compute_result(mut cx: FunctionContext) -> JsResult<JsObject> {
    let js_info = cx.argument::<JsObject>(0)?;
    promise(&mut cx, |cx, future_context| {
        let saved_info = future_context.register_context_data(cx, js_info);
        let trait_impl = TraitImpl::new(saved_info, future_context);
        async move {
            let result = compute_result(trait_impl).await?;
            fulfill_promise(move |cx| Ok(cx.string(result)))
        }
    })
}